pyproject.toml
[project]
name = "mcp-a2a-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.24",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
